+++ tests/read_test.rs
@@ -4,8 +4,9 @@ use std::io::{BufReader, Read, Seek, SeekFrom};
 use exif::parse_exif;
 
 use libheif_rs::{
-    check_file_type, Chroma, ColorSpace, CompressionFormat, EncoderParameterValue, EncoderQuality,
-    FileTypeResult, HeifContext, ItemId, Result, RgbChroma, StreamReader,
+    check_file_type, color_profile_types, Chroma, ColorPrimaries, ColorProfile, ColorSpace,
+    CompressionFormat, EncoderParameterValue, EncoderQuality, FileTypeResult, HeifContext, ItemId,
+    MatrixCoefficients, Result, RgbChroma, StreamReader, TransferCharacteristics,
 };
 
 #[test]
@@ -22,7 +23,7 @@ fn read_from_file() -> Result<()> {
 fn read_from_reader() -> Result<()> {
     let mut file = BufReader::new(File::open("./data/test.heic").unwrap());
     let total_size = file.seek(SeekFrom::End(0)).unwrap();
-    file.seek(SeekFrom::Start(0)).unwrap();
+    file.rewind().unwrap();
     let stream_reader = StreamReader::new(file, total_size);
 
     let ctx = HeifContext::read_from_reader(Box::new(stream_reader))?;
@@ -30,7 +31,7 @@ fn read_from_reader() -> Result<()> {
     assert_eq!(handle.width(), 3024);
     assert_eq!(handle.height(), 4032);
 
-    let src_img = handle.decode(ColorSpace::Undefined, false)?;
+    let src_img = handle.decode(ColorSpace::Undefined, None)?;
     assert_eq!(
         src_img.color_space(),
         Some(ColorSpace::Rgb(RgbChroma::C444))
@@ -119,7 +120,7 @@ fn decode_and_scale_image() -> Result<()> {
     let handle = ctx.primary_image_handle()?;
 
     // Decode the image
-    let src_img = handle.decode(ColorSpace::YCbCr(Chroma::C420), false)?;
+    let src_img = handle.decode(ColorSpace::YCbCr(Chroma::C420), None)?;
     assert_eq!(src_img.color_space(), Some(ColorSpace::YCbCr(Chroma::C420)));
     let planes = src_img.planes();
     let y_plane = planes.y.unwrap();
@@ -234,4 +235,86 @@ fn test_check_file_type() {
     assert_eq!(check_file_type(&data[..7]), FileTypeResult::MayBe);
     assert_eq!(check_file_type(&data[..8]), FileTypeResult::MayBe);
     assert_eq!(check_file_type(&data[..12]), FileTypeResult::Supported);
+}
+
+#[test]
+fn test_raw_color_profile_of_image_handle() -> Result<()> {
+    let ctx = HeifContext::read_from_file("./data/test.heic")?;
+    let handle = ctx.primary_image_handle()?;
+
+    let raw_profile = handle.color_profile_raw().unwrap();
+    assert_eq!(raw_profile.profile_type(), color_profile_types::PROF);
+    assert_eq!(raw_profile.data.len(), 548);
+
+    let nclx_profile = handle.color_profile_nclx();
+    assert!(nclx_profile.is_none());
+    Ok(())
+}
+
+#[test]
+fn test_nclx_color_profile_of_image_handle() -> Result<()> {
+    let ctx = HeifContext::read_from_file("./data/test_nclx.heif")?;
+    let handle = ctx.primary_image_handle()?;
+
+    let raw_profile = handle.color_profile_raw();
+    assert!(raw_profile.is_none());
+
+    let nclx_profile = handle.color_profile_nclx().unwrap();
+    assert_eq!(nclx_profile.profile_type(), color_profile_types::NCLX);
+    assert_eq!(
+        nclx_profile.color_primaries(),
+        ColorPrimaries::ITU_R_BT_2020_2_and_2100_0
+    );
+    assert_eq!(
+        nclx_profile.transfer_characteristics(),
+        TransferCharacteristics::ITU_R_BT_2100_0_PQ
+    );
+    assert_eq!(
+        nclx_profile.matrix_coefficients(),
+        MatrixCoefficients::RGB_GBR
+    );
+    assert_eq!(nclx_profile.full_range_flag(), 1);
+    assert_eq!(nclx_profile.color_primary_red_x(), 0.708);
+    assert_eq!(nclx_profile.color_primary_red_y(), 0.292);
+    assert_eq!(nclx_profile.color_primary_green_x(), 0.17);
+    assert_eq!(nclx_profile.color_primary_green_y(), 0.797);
+    assert_eq!(nclx_profile.color_primary_blue_x(), 0.131);
+    assert_eq!(nclx_profile.color_primary_blue_y(), 0.046);
+    assert_eq!(nclx_profile.color_primary_white_x(), 0.3127);
+    assert_eq!(nclx_profile.color_primary_white_y(), 0.329);
+    Ok(())
+}
+
+#[test]
+fn test_raw_color_profile_of_image() -> Result<()> {
+    let ctx = HeifContext::read_from_file("./data/test.heic")?;
+    let handle = ctx.primary_image_handle()?;
+    let image = handle.decode(ColorSpace::Undefined, None)?;
+    // Hmm... It's strange. Decoded image doesn't have color profile.
+    assert!(image.color_profile_raw().is_none());
+    assert!(image.color_profile_nclx().is_none());
+    Ok(())
+}
+
+#[test]
+fn test_read_avif_image() -> Result<()> {
+    let ctx = HeifContext::read_from_file("./data/test_nclx.avif")?;
+    let handle = ctx.primary_image_handle()?;
+
+    assert_eq!(handle.width(), 2048);
+    assert_eq!(handle.height(), 1440);
+    assert_eq!(handle.luma_bits_per_pixel(), 12);
+    assert_eq!(handle.chroma_bits_per_pixel(), 12);
+
+    let nclx_profile = handle.color_profile_nclx();
+    assert!(nclx_profile.is_some());
+
+    let image = handle.decode(ColorSpace::Undefined, None)?;
+    assert_eq!(image.color_space(), Some(ColorSpace::Rgb(RgbChroma::C444)));
+    let planes = image.planes();
+    let r_plane = planes.r.unwrap();
+    assert_eq!(r_plane.width, 2048);
+    assert_eq!(r_plane.height, 1440);
+
+    Ok(())
 }